+++ src/y0/algorithm/counterfactual_transport/__init__.py
@@ -1,0 +1,29 @@
+# -*- coding: utf-8 -*-
+
+"""Implementation of counterfactual transportability.
+
+.. [correa22a] https://proceedings.mlr.press/v162/correa22a/correa22a.pdf.
+"""
+
+from .api import (
+    CFTDomain,
+    ConditionalCFTResult,
+    Event,
+    UnconditionalCFTResult,
+    conditional_cft,
+    transport_conditional_counterfactual_query,
+    transport_unconditional_counterfactual_query,
+    unconditional_cft,
+)
+
+__all__ = [
+    "unconditional_cft",
+    "conditional_cft",
+    "transport_unconditional_counterfactual_query",
+    "transport_conditional_counterfactual_query",
+    #
+    "Event",
+    "CFTDomain",
+    "ConditionalCFTResult",
+    "UnconditionalCFTResult",
+]
